requirements.txt
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
backoff==2.2.1
beautifulsoup4==4.13.3
browser-use==0.1.40
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
colorlog==6.9.0
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
graphql-core==3.2.6
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-anthropic==0.3.3
langchain-core==0.3.51
langchain-ollama==0.2.2
langchain-openai==0.3.1
langsmith==0.3.25
markdown-it-py==3.0.0
markdownify==0.14.1
MarkupSafe==3.0.2
mdurl==0.1.2
monotonic==1.6
ollama==0.4.7
openai==1.71.0
orjson==3.10.16
packaging==24.2
playwright==1.51.0
posthog==3.23.0
pydantic==2.11.2
pydantic_core==2.33.1
pyee==12.1.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.1
setuptools==78.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.46.1
strawberry-graphql==0.270.2
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.5
websockets==15.0.1
zstandard==0.23.0
